refactor(cli): extract command_catalog tests into sibling module (#609 phase 1)
Split the largest god module in heddle#609 — `command_catalog.rs`
(6777 LOC) — by moving its 1993-line `#[cfg(test)] mod tests` block
into a sibling `command_catalog/tests.rs`. The former single file
becomes a `command_catalog/` directory:

  command_catalog.rs            (6777 LOC)
  -> command_catalog/mod.rs     (4785 LOC, generator + types + contracts)
  -> command_catalog/tests.rs   (2000 LOC, the test module, verbatim)

Pure, behavior-preserving re-modularization: the test body moved
unchanged (de-indented one level out of the `mod tests {}` wrapper)
and still binds to the parent via `super::*`, exactly as it did
inline. No public item renamed, no signature changed, no logic
touched — the parent module only gains a `#[cfg(test)] mod tests;`
declaration in place of the inline block. The catalog's public
surface (`pub use command_catalog::{...}` in commands/mod.rs) is
untouched.

This is phase 1 of the #609 god-module decomposition; remaining
clusters are tracked as follow-up phases in the PR body.

Gates: cargo build --locked --workspace (default + --all-features),
check-no-silent-default-tree-load.sh, clippy --workspace
--all-targets -D warnings, command_catalog lib tests (30 pass).
